changed minimal raise type

+++ models/offer.go
@@ -1,9 +1,9 @@
 package models
 
 type Offer struct {
-	Sender string  `bson:"sender" json:"sender"`
-	Price  float32 `bson:"offer" json:"offer"`
-	Time   int64   `bson:"time" json:"time"`
+	Sender string `bson:"sender" json:"sender"`
+	Price  int64  `bson:"offer" json:"offer"`
+	Time   int64  `bson:"time" json:"time"`
 }
 type OfferUnwind struct {
 	Offers Offer `json:"offers" bson:"offers"`
